Change workflow trigger from tag to branch in main.yml
The commit modifies the GitHub actions workflow main.yml to trigger on a push event to 'main' branch instead of 'v' tag. This prevents unnecessary actions being triggered when making a release or pushing a tag. For standardization, it ensures that the pipeline runs and tests are performed on new code pushed to the 'main' branch instead of waiting for a tag to be added, improving automation and CI/CD practices.