Update Workflow Triggers in main.yml
This commit adjusts the GitHub Actions workflow main.yml to trigger on push events to the 'main' branch as opposed to 'v' prefixed tags. The update aims to enhance automation and continuous integration/deployment practices by ensuring pipeline runs and tests are performed on new code pushed to 'main', instead of waiting for a tag addition. This prevents the triggering of unnecessary actions during releases or tag pushes.